Add set price and get price to ItemCatalog for items priced by weight. Refactor doubles to big decimals for precision in multiplication and rounding.

+++ src/main/java/com/kpmac/checkout/Item.java
@@ -1,12 +1,14 @@
 package com.kpmac.checkout;
 
+import java.math.BigDecimal;
+
 public class Item {
 
     private final Boolean pricedByWeight;
     private final String name;
-    private final Double price;
+    private final BigDecimal price;
 
-    public Item(String name, Double price, Boolean pricedByWeight) {
+    public Item(String name, BigDecimal price, Boolean pricedByWeight) {
         this.name = name;
         this.price = price;
         this.pricedByWeight = pricedByWeight;
@@ -16,7 +18,7 @@ public class Item {
         return pricedByWeight;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
